+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * New functions `pdf_crop()`, `write_datasets()`, `attsurf2df()`.
 
+* Fix bug that made polygons extend beyond their boundaries affecting
+  `eafdiff()` and `eafdiffplot()`.
+
 
 # eaf 2.2